Respond to reviewer comments
