+++ executable_semantics/BUILD
@@ -13,6 +13,8 @@ cc_binary(
     name = "executable_semantics",
     srcs = ["main.cpp"],
     deps = [
+        "//executable_semantics/common:arena",
+        "//executable_semantics/common:nonnull",
         "//executable_semantics/interpreter:exec_program",
         "//executable_semantics/syntax",
         "@llvm-project//llvm:Support",
