+++ executable_semantics/syntax/BUILD
@@ -4,7 +4,7 @@
 
 load("@mypy_integration//:mypy.bzl", "mypy_test")
 
-package(default_visibility = ["//executable_semantics:__subpackages__"])
+package(default_visibility = ["//executable_semantics:__pkg__"])
 
 cc_library(
     name = "bison_wrap",
